requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Quantum computing
pennylane>=0.33.0
pennylane-qiskit>=0.33.0
qiskit>=0.45.0
qiskit-aer>=0.13.0
qiskit-ibm-runtime>=0.15.0

# Protein structure
biopython>=1.81
prody>=2.4.0

# Deep learning utilities
transformers>=4.35.0
fair-esm>=2.0.0
torch-geometric>=2.4.0

# Symbolic reasoning
pyswip>=0.2.10  # Prolog interface
z3-solver>=4.12.0  # SMT solver

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
py3Dmol>=2.0.0
plotly>=5.17.0

# Utilities
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0
jupyter>=1.0.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Formal verification (optional)
# lean4>=4.0.0  # Install separately via elan
